feat: align tables left, handle cutting/bend formats, restrict lorry number length, hide manager edit final button, update location staff Lots Allotted tab, fix AllottedSupervisors user ReferenceError

+++ server/__tests__/directLoadLocationHandling.property.test.js
@@ -156,7 +156,7 @@ describe('Property Test 9: Direct Load Location Handling', () => {
           outturnId: fc.constantFrom(...testOutturns.map(o => o.id)),
           productType: fc.constant('Rice'),
           requestedBags: fc.integer({ min: 1, max: 10 }),
-          requestedQtls: fc.float({ min: 0.1, max: 5.0 }),
+          requestedQtls: fc.float({ min: Math.fround(0.1), max: Math.fround(5.0)  }),
           date: fc.constant(new Date().toISOString().split('T')[0])
         }),
         async (testData) => {
@@ -282,7 +282,7 @@ describe('Property Test 9: Direct Load Location Handling', () => {
           outturnId: fc.constantFrom(...testOutturns.map(o => o.id)),
           productType: fc.constant('Rice'),
           bags: fc.integer({ min: 1, max: 5 }),
-          quantityQuintals: fc.float({ min: 0.1, max: 2.0 }),
+          quantityQuintals: fc.float({ min: Math.fround(0.1), max: Math.fround(2.0)  }),
           date: fc.constant(new Date().toISOString().split('T')[0])
         }),
         async (testData) => {
